enable an experimental reversal of the julia main loop
this makes Gtk the main event-poll function

this also uses a slower poll of the julia loop than before.

assuming that most user interaction will be through the gui,
the speed that external events are reacted to is less important

if there is a uv_backend_fd (available on most unix systems),
we use that too, so there is no impact on the reaction speed to external
events. but still we still do a slow poll of the libuv main loop (since
some backends will not emit event notifications for certain types of
events -- see uv.h for more details)